% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/addMemoization.R
\name{addMemoization}
\alias{addMemoization}
\title{Add Memoization to a Function}
\usage{
addMemoization(fcn)
}
\arguments{
\item{fcn}{the function to be memoized}
}
\value{
memoized function
}
\description{
Add Memoization to a Function
}

