% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setCacheRootPath.R
\name{setCacheRootPath}
\alias{setCacheRootPath}
\title{Set Cache Root Path}
\usage{
setCacheRootPath(path = "~/.simpleRCache")
}
\arguments{
\item{path}{the cache root directory}
}
\description{
Set Cache Root Path
}
\examples{
setCacheRootPath()

}

