% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMemoization.R
\name{addMemoization}
\alias{addMemoization}
\title{Add Memoization to a Function}
\usage{
addMemoization(fcn, verbose = FALSE)
}
\arguments{
\item{fcn}{the function to be memoized}

\item{verbose}{show debugging information}
}
\value{
memoized function
}
\description{
Add Memoization to a Function
}
\examples{
runifCached <- addMemoization(runif)

}
