% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setCacheRootPath.R
\name{setCacheRootPath}
\alias{setCacheRootPath}
\title{Set Cache Root Path}
\usage{
setCacheRootPath(path = tempdir())
}
\arguments{
\item{path}{the cache root directory (DEFAULT: a temporary directory, 
but this can be change to a more permanent location)}
}
\description{
Set Cache Root Path
}
\examples{
setCacheRootPath()
\dontrun{
setCacheRootPath("~/.simpleRCache")
}

}
